retransmit after timeout done
also fixed rule processing so that it would run before multicast
messages processed. had to comment out a dupe if statement that was
keeping rules from processing. May need to change structure of sequence
numbers as the multicast from Alice has the same sequence number as the
first multicast from Bob, which could be a problem for matching Acks.